# docs/api/recovery-control-plane-openapi.yaml
openapi: 3.1.0
...
info:
  title: Windows Recovery MVP Control Plane API
  version: 1.0.0
  summary: Recovery MVP v1 的单实例 Control Plane 权威线上合同
  description: |
    Control Plane 接收 Agent 主动注册/心跳，维护服务镜像，并按严格依赖 DAG 调度
    RecoveryRun。Agent ingress 只接受实验 `cluster_token`；管理 API 只接受管理员签名
    Session，且所有管理写请求同时要求 CSRF Header。两类凭据不可互换。

    HTTP 明文模式只允许实验或受控管理网。Control Plane 必须部署在独立管理节点，
    其 SQLite 不得依赖被管理的 MySQL、Redis 或 Nacos。
servers:
  - url: http://{control_plane_ip}:{control_plane_port}
    description: 单实例 Windows Control Plane
    variables:
      control_plane_ip:
        default: 127.0.0.1
      control_plane_port:
        default: "8766"
...
security:
  - adminSession: []
paths:
...
  /api/v1/agents/register:
    post:
      tags: [Agent ingress]
      operationId: registerAgent
      summary: 注册 Agent 当前进程实例
      description: |
        只接受比已存 generation 更大的实例，或完全相同 `(instance_generation,
        agent_instance_id)` 的重试。完全相同实例仍执行 sequence 规则：sequence 不大于
        last_sequence 时幂等忽略，且不刷新 lease、endpoint 或服务镜像。更小 generation 或
        同 generation 不同 instance 返回 `409 STALE_AGENT_INSTANCE`。CP 只接受无 userinfo/path/query/fragment 的
        `http://IP-literal:port`，以 socket peer IP 和已校验端口重建 endpoint，不读取
        `X-Forwarded-For`。
      security:
        - clusterToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentReport"
            example:
              agent_id: b8d82b70-5a8a-4c07-8f44-31ef85a6aa01
              boot_id: dba11591-4a51-49bb-8cf0-7e3c2e85e517
              agent_instance_id: a82a5615-dcf8-4f8a-a266-dfe935f83b27
              instance_generation: 7
              sequence: 1
              version: 0.1.0
              endpoint: http://10.20.0.11:8765
              hostname: APP-SRV-01
              services:
                - local_service_id: mysql
                  windows_service_name: MySQL80
                  display_name: MySQL 8
                  installation_state: INSTALLED
                  runtime_state: INACTIVE
                  startup_state: AUTOSTART_DISABLED
                  last_observed_at: "2026-07-16T08:00:00Z"
      responses:
        "200":
          description: 注册已接受；完全相同实例的重试也返回 200。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeartbeatAck"
              example:
                accepted: true
                ignored: false
                reason: ACCEPTED
                received_at: "2026-07-16T08:00:01Z"
                lease_updated_at: "2026-07-16T08:00:01Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/agents/{agent_id}/heartbeat:
    post:
      tags: [Agent ingress]
      operationId: ingestHeartbeat
      summary: 接收已注册实例的主动心跳
      description: |
        路径与正文 agent_id 必须相同，且 `(agent_instance_id, instance_generation)` 必须为
        当前已注册实例。旧实例返回 `STALE_AGENT_INSTANCE`。同一实例中
        `sequence <= last_sequence` 的报告返回 200 并忽略，且不刷新 received_at 或服务镜像。
      security:
        - clusterToken: []
      parameters:
        - $ref: "#/components/parameters/AgentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentReport"
      responses:
        "200":
          description: 心跳已接受，或作为重复/乱序报告幂等忽略。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeartbeatAck"
              examples:
                accepted:
                  value:
                    accepted: true
                    ignored: false
                    reason: ACCEPTED
                    received_at: "2026-07-16T08:00:11Z"
                    lease_updated_at: "2026-07-16T08:00:11Z"
                ignored:
                  value:
                    accepted: false
                    ignored: true
                    reason: DUPLICATE_OR_OUT_OF_ORDER
                    received_at: "2026-07-16T08:00:12Z"
                    lease_updated_at: null
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/agents:
    get:
      tags: [Inventory]
      operationId: listAgents
      summary: 列出 Agent 在线租约
      description: "`ONLINE/OFFLINE` 仅由 CP 接收时间和 45 秒 lease timeout 计算。"
      responses:
        "200":
          description: Agent 当前租约。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentCollection"
              example:
                items:
                  - agent_id: b8d82b70-5a8a-4c07-8f44-31ef85a6aa01
                    boot_id: dba11591-4a51-49bb-8cf0-7e3c2e85e517
                    agent_instance_id: a82a5615-dcf8-4f8a-a266-dfe935f83b27
                    instance_generation: 7
                    hostname: APP-SRV-01
                    endpoint: http://10.20.0.11:8765
                    version: 0.1.0
                    online_status: ONLINE
                    last_sequence: 18
                    last_received_at: "2026-07-16T08:03:01Z"
                    offline_after_seconds: 45
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/services:
    get:
      tags: [Inventory]
      operationId: listManagedServices
      summary: 列出 CP 服务镜像
      responses:
        "200":
          description: 每项以 CP 分配的 `managed_service_id` 标识。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedServiceCollection"
              example:
                items:
                  - managed_service_id: 72961a6e-35cd-46c8-b405-56d73dbf3619
                    agent_id: b8d82b70-5a8a-4c07-8f44-31ef85a6aa01
                    agent_online_status: ONLINE
                    local_service_id: mysql
                    windows_service_name: MySQL80
                    display_name: MySQL 8
                    installation_state: INSTALLED
                    runtime_state: ACTIVE
                    startup_state: AUTOSTART_DISABLED
                    last_observed_at: "2026-07-16T08:00:00Z"
                    mirrored_at: "2026-07-16T08:00:01Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/services/{managed_service_id}/actions/{action}:
    post:
      tags: [Service actions]
      operationId: proxyServiceAction
      summary: 代理单服务 start、stop 或 restart
      description: |
        CP 将 managed_service_id 映射为所属 Agent 的 local_service_id。涉及非终态 RecoveryRun
        独占的服务返回 `409 SERVICE_IN_ACTIVE_RUN`。Idempotency-Key 原样传给 Agent。正文可省略；
        若提供只能是严格空对象 `{}`。
      security:
        - adminSession: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/ManagedServiceId"
        - $ref: "#/components/parameters/ServiceAction"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmptyActionRequest"
            example: {}
      responses:
        "202":
          description: Agent Operation 已创建或命中幂等记录。
          headers:
            Location:
              $ref: "#/components/headers/OperationLocation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxiedOperation"
              example:
                managed_service_id: 72961a6e-35cd-46c8-b405-56d73dbf3619
                operation_id: 1e881030-2451-4d7f-9628-6a58c2912e14
                agent_id: b8d82b70-5a8a-4c07-8f44-31ef85a6aa01
                local_service_id: mysql
                windows_service_name: MySQL80
                action: start
                status: PENDING
                idempotency_key: db7df82f-c129-4e7a-bc51-6bcd4433fe74
                request_fingerprint: 7bc1e2673f2333ac92a479a3aa55e91fc19451269d5c2f33c3423178b00026e8
                created_at: "2026-07-16T08:00:02Z"
                updated_at: "2026-07-16T08:00:02Z"
                started_at: null
                finished_at: null
                error_code: null
                message: null
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/operations/{operation_id}:
    get:
      tags: [Service actions]
      operationId: getProxiedOperation
      summary: 查询 CP 代理创建的 Agent Operation
      parameters:
        - $ref: "#/components/parameters/OperationId"
      responses:
        "200":
          description: Operation 当前状态；UNKNOWN 不得自动重放。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxiedOperation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/recovery-groups:
    get:
      tags: [Recovery groups]
      operationId: listRecoveryGroups
      summary: 列出恢复组
      responses:
        "200":
          description: 恢复组及其当前触发状态。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryGroupCollection"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Recovery groups]
      operationId: createRecoveryGroup
      summary: 创建未 arm 的空恢复组
      security:
        - adminSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryGroupCreate"
            example:
              name: core-stack
              description: MySQL and Redis to Nacos to Java to Nginx
              node_settle_window_seconds: 120
              max_parallel_services: 4
      responses:
        "201":
          description: 恢复组已创建。
          headers:
            Location:
              $ref: "#/components/headers/RecoveryGroupLocation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryGroup"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
  /api/v1/recovery-groups/{group_id}/arm:
    post:
      tags: [Recovery groups]
      operationId: armRecoveryGroup
      summary: 校验恢复组并记录当前 epoch 基线
      description: |
        DISARMED 时，所有 required Agent 必须在线，服务必须 INSTALLED 且
        AUTOSTART_DISABLED，图必须无环；成功仅保存 baseline_epoch，不创建 AUTO Run、不启动服务。
        其他任意状态重复调用均幂等返回当前组，不重置 baseline、candidate、settle 计时或 Run。
        BLOCKED_PRECONDITION 必须显式 disarm、修复后重新 arm 才能解除。
      security:
        - adminSession: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/GroupId"
      responses:
        "200":
          description: 首次 arm 后进入 ARMED_IDLE；已 arm 时返回未改变的当前组。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryGroup"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    clusterToken:
      type: http
      scheme: bearer
      bearerFormat: opaque-cluster-token
      description: 仅用于 Agent 注册/心跳；至少 32 个随机字节并常量时间比较。
    adminSession:
      type: apiKey
      in: cookie
      name: recovery_admin_session
      description: 单管理员登录后签发的 HttpOnly、SameSite 签名 Session Cookie。
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: 所有管理 API 写请求必须与管理员 Session 中的 CSRF 值匹配。
...
  responses:
    Unauthorized:
      description: 对当前接口缺少正确的 cluster token 或管理员 Session。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: AUTH_INVALID
            message: Authentication credential is missing or invalid
            detail: null
            request_id: 7cf653dc-3bba-428e-8d5e-1bfd9220a60f
...
  schemas:
...
    EmptyActionRequest:
      type: object
      maxProperties: 0
      additionalProperties: false
...
    HeartbeatAck:
      type: object
      additionalProperties: false
      required: [accepted, ignored, reason, received_at, lease_updated_at]
      properties:
        accepted:
          type: boolean
        ignored:
          type: boolean
        reason:
          type: string
          enum: [ACCEPTED, DUPLICATE_OR_OUT_OF_ORDER]
        received_at:
          type: string
          format: date-time
        lease_updated_at:
          type: [string, "null"]
          format: date-time
...